Cast to intptr_t instead of int to kick out the final warning

+++ python/rpmmodule.c
@@ -491,6 +491,6 @@ void init_rpm(void)
 
     REGISTER_ENUM(RPMDBI_PACKAGES);
 
-    REGISTER_ENUM((int)RPMAL_NOMATCH);
+    REGISTER_ENUM((intptr_t)RPMAL_NOMATCH);
 }
